requirements.txt
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
autoflake==2.2.1
autopep8==2.0.4
backcall==0.2.0
black==23.9.1
certifi==2023.7.22
chirho==0.1.0
-e git+https://github.com/BasisResearch/cities.git@837106c92e7c329966d28ec4526aa627e59f6b13#egg=cities
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.4
contourpy==1.1.1
coverage==7.3.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
execnet==2.0.2
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
fiona==1.9.5
flake8==6.1.0
fonttools==4.43.1
geopandas==0.14.0
graphviz==0.20.1
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
mypy==1.6.0
mypy-extensions==1.0.0
nbformat==5.9.2
nbqa==1.7.1
nbval==0.10.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.17.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.6.1
pyro-api==0.1.2
pyro-ppl==1.8.6
pytest==7.4.2
pytest-xdist==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
rpds-py==0.10.4
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenize-rt==5.2.0
tomli==2.0.1
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.8
